+++ main.tf
@@ -30,6 +30,7 @@ locals {
   app                   = lower(try(length(local.pulsar_app_name), 0) > 0 ? local.pulsar_app_name : local.workspace_name)
   db_name               = replace(local.app, "-", "_")
   dev_url               = "https://webapp--main--${local.workspace_name}--${local.user_username}.embold.dev"
+  dotfiles_url          = data.coder_parameter.dotfiles_url.value
   mariadb_version       = data.coder_parameter.mariadb_version.value
   mariadb_auto_upgrade  = data.coder_parameter.mariadb_auto_upgrade.value ? "1" : "0"
   php_version           = data.coder_parameter.php_version.value
@@ -46,6 +47,12 @@ locals {
 
 variable "DOCKER_REGISTRY_PASS" {
   sensitive = true
+}
+
+data "coder_parameter" "dotfiles_url" {
+  name        = "dotfiles URL"
+  description = "GitHub repository with dotfiles"
+  mutable     = true
 }
 
 data "coder_parameter" "pulsar_app_name" {
@@ -141,6 +148,7 @@ resource "coder_agent" "main" {
     CODER_WORKSPACE_NAME  = local.workspace_name
     CODER_WORKSPACE_PORT  = 443
     DEVURL                = local.dev_url
+    DOTFILES_URL          = local.dotfiles_url
     GIT_AUTHOR_NAME       = local.user_full_name
     GIT_AUTHOR_EMAIL      = local.user_email
     GIT_COMMITTER_NAME    = local.user_full_name